add squad-compare-builds
This scripts makes it possible to compare builds within a group/project
or between different groups or projects.

Requires squad-client PR-162 [1], checkout this PR and do 'pip3 install
-e .' in order to use this new 'download-results' feature from
squad-client.

To being able to say latest, latest-1 or even v5.19.4+1, code was
borrowed. Function get_build was copied from PR-162 for now, since its
not a util in squad-client. When it will be a util in squad-client this
code will be changed to use that function instead.

Usage:
./squad-compare-builds --group lkft --project linux-stable-rc-linux-5.19.y --build latest-1,latest
./squad-compare-builds --group lkft --project linux-mainline-master,linux-stable-rc-linux-5.19.y --build latest-1,latest

[1] https://github.com/Linaro/squad-client/pull/162

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>