squadutilslib: Add get_reproducer_from_testrun
Add get_reproducer_from_testrun to allow downloading and returning of a
reproducer from a given testrun ID. This function downloads the
reproducer for the given testrun to a provided filename. The optional
`local` flag can be provided to download the tuxmake or tuxrun
reproducer command rather than the tuxsuite reproducer command.

This function will be used in a script to download the reproducer file
for a build or test from a TestRun, which is useful for reproducing
builds and tests.

Update get_reproducer to use the get_reproducer_from_testrun function to
download the reproducer from the testrun.

Signed-off-by: Katie Worton <katie.worton@linaro.org>