.github: Fix flake8 command to check non-.py Python files
Some of the scripts in the repo are Python files that flake8 should
check, but flake8 only checks .py files. Additionally, flake8 may crash
if given a non-Python file. Work around these issues by passing
"squad-*" to flake8, so it checks the Python scripts that begin with
"squad-" from this repo.

Fixes: c4f195c (".github: enable flake8 checks")
Signed-off-by: Katie Worton <katie.worton@linaro.org>